FIX: Windows build
This contains a series of fixes + changes to make sure we
can build glib as a dll using bazel.

First, glib is mixing headers + c files, something bazel doesn't really
like. So we rename all .c files that are actually treated as headers.

Next we introduce a rule to create windows dll's, which we have taken
from the google public repo.

We also bring in the tests.

Change-Id: I37b9946abb2e94e15d0ae45ca836b165a778e7ce