Split glib into shared and static libs
The static library relies on static constructors being run when a the
library is used. These constructors initialize a set of global locks,
etc.  Without these initializers you can see unexpected crashes.

This behavior is triggered through a define that should be passed in
during compilation, and inclusion by dependents.

Since the behavior between the static and shared library is different we
now split the library in a shared and static version and specify the
needed defines.

Bug: 311275700
Change-Id: I2ff87341fc9709c5b504d199b5d08ca2939d96ea